requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot1/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
anyio==4.1.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot1/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
h11==0.14.0
httpcore==1.0.2
httpx==0.27.2
idna==3.6
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot1/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
mock-open==1.4.0
pyTelegramBotAPI==4.14.0
python-telegram-bot==21.7
pytz==2023.3.post1
requests==2.31.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot1/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
tornado==6.4
tzlocal==5.2
urllib3==2.1.0
